tradesage/requirements.txt
# Core
fastapi>=0.115
uvicorn[standard]>=0.34
websockets>=13
python-dotenv>=1.0
pydantic>=2.10
pydantic-settings>=2.7

# LangChain + LLM
langchain>=0.3
langchain-anthropic>=0.3
langchain-community>=0.3
langchain-core>=0.3
langgraph>=0.2
anthropic>=0.49

# Knowledge Graph
neo4j>=5.26
langchain-neo4j>=0.3

# Data
alpaca-py>=0.38
ccxt>=4.4
yfinance>=0.2
tavily-python>=0.5

# Statistics
scipy>=1.15
numpy>=2.2
pandas>=2.2

# Database
sqlalchemy>=2.0
aiosqlite>=0.20

# PDF processing
pymupdf>=1.25

# Scheduling
apscheduler>=3.11

# HTTP
httpx>=0.28
aiohttp>=3.11
requests>=2.32
tenacity>=9.0

# Utilities
python-multipart>=0.0.20
python-jose>=3.3

# Vector embeddings
openai>=1.59
